source-loader: separate server and client code
