+++ .babelrc.js
@@ -54,7 +54,6 @@ module.exports = {
     },
     {
       test: './lib',
-      exclude: './addons/storysource/src/loader',
       presets: [
         ['@babel/preset-env', { shippedProposals: true, useBuiltIns: 'usage', corejs: '3' }],
         '@babel/preset-react',
@@ -75,14 +74,11 @@ module.exports = {
     },
     {
       test: [
-        './lib/core/src/server',
         './lib/node-logger',
         './lib/codemod',
-        './lib/source-loader/src',
         './addons/storyshots',
-        './addons/storysource/src/loader',
-        './app/**/src/server/**',
-        './app/**/src/bin/**',
+        '**/src/server/**',
+        '**/src/bin/**',
       ],
       presets: [
         [
