+++ lib/source-loader/package.json
@@ -16,8 +16,8 @@
     "directory": "lib/source-loader"
   },
   "license": "MIT",
-  "main": "dist/index.js",
-  "jsnext:main": "src/index.js",
+  "main": "dist/server/index.js",
+  "jsnext:main": "src/server/index.js",
   "scripts": {
     "prepare": "node ../../scripts/prepare.js"
   },
