+++ lib/source-loader/src/client/events.js
