+++ lib/source-loader/src/client/preview.js
