--- lib/source-loader/src/index.js
@@ -1,6 +1,0 @@
-import { transform } from './build';
-
-export { STORY_EVENT_ID } from './events';
-export { addSource, withSource } from './preview';
-
-export default transform;
