+++ lib/source-loader/src/server/abstract-syntax-tree/default-options.js
