+++ lib/source-loader/src/server/abstract-syntax-tree/generate-helpers.js
@@ -68,7 +68,7 @@ const STORY_DECORATOR_STATEMENT =
 const ADD_PARAMETERS_STATEMENT =
   '.addParameters({ storySource: { source: __STORY__, locationsMap: __ADDS_MAP__ } })';
 const IMPORT_DECLARATION_FOR_EXPORTED_STORIES_DECORATOR =
-  'var addSourceDecorator = require("@storybook/source-loader/dist/preview").addSource;\n';
+  'var addSourceDecorator = require("@storybook/source-loader/preview").addSource;\n';
 const applyExportDecoratorStatement = part =>
   ` addSourceDecorator(${part}, {__STORY__, __ADDS_MAP__,__MAIN_FILE_LOCATION__,__MODULE_DEPENDENCIES__,__LOCAL_DEPENDENCIES__,__SOURCE_PREFIX__,__IDS_TO_FRAMEWORKS__});`;
