+++ lib/source-loader/src/server/abstract-syntax-tree/inject-decorator.test.js
