+++ lib/source-loader/src/server/abstract-syntax-tree/parse-helpers.js
