+++ lib/source-loader/src/server/build.js
@@ -15,7 +15,7 @@ export function transform(inputSource) {
         localDependencies,
         idsToFrameworks,
       }) => `
-  var withSourceLoader = require('@storybook/source-loader/dist/preview').withSource;
+  var withSourceLoader = require('@storybook/source-loader/preview').withSource;
   var __SOURCE_PREFIX__ = "${prefix.replace(/\\([^\\ ])/g, '\\\\$1')}";
   var __STORY__ = ${sourceJson};
   var __ADDS_MAP__ = ${JSON.stringify(addsMap)};
