+++ lib/source-loader/src/server/dependencies-lookup/getRidOfUselessFilePrefixes.js
