+++ lib/source-loader/src/server/dependencies-lookup/readAsObject.js
