+++ lib/source-loader/src/server/index.js
@@ -1,0 +1,3 @@
+import { transform } from './build';
+
+export default transform;
